Fix compilation with msys2

+++ src/action.c
@@ -53,13 +53,13 @@ static long get_arg_value(arg_t arg, const arg_t *args)
     }
     // Default values for some types.
     if (arg.type == TYPE_GOXEL)
-        return (long)goxel();
+        return (intptr_t)goxel();
     if (arg.type == TYPE_IMAGE)
-        return (long)goxel()->image;
+        return (intptr_t)goxel()->image;
     if (arg.type == TYPE_LAYER)
-        return (long)goxel()->image->active_layer;
+        return (intptr_t)goxel()->image->active_layer;
     if (arg.type == TYPE_BOX)
-        return (long)&goxel()->selection;
+        return (intptr_t)&goxel()->selection;
     if (str_equ(arg.name, "width"))
         return (int)goxel()->image->export_width;
     if (str_equ(arg.name, "height"))
